requirements.txt
discord.py=1.7.3
ffmpeg=1.4
python-dotenv=0.20.0
youtube-dl=2021.12.17
requests=2.27.1
